working sysmem and panic independence

+++ kernel/memory/sysmem.h
@@ -2,8 +2,8 @@
 #include <stddef.h>
 #include <stdint.h>
 typedef struct sysmem_segment {
-    uint32_t size;
+    size_t size;
     sysmem_segment* next;
     sysmem_segment* prev;
 } sysmem_t;
-void init_sysmem(uint32_t start, uint32_t size);
+void init_sysmem(uint64_t start, size_t size);
